De-Duplication - 1) Reused repetitive code by extracting to methods 2) Pulled common code to common class and inherited the class
