+++ src/main/java/org/recap/util/CommonUtil.java
@@ -1,19 +1,42 @@
 package org.recap.util;
 
 import org.apache.commons.lang3.StringUtils;
+import org.marc4j.marc.Record;
+import org.recap.RecapCommonConstants;
 import org.recap.RecapConstants;
-import org.recap.model.jpa.BibliographicEntity;
-import org.recap.model.jpa.HoldingsEntity;
-import org.recap.model.jpa.ItemEntity;
+import org.recap.model.jpa.*;
 import org.recap.model.report.SubmitCollectionReportInfo;
+import org.recap.repository.jpa.*;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.Date;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 @Service
 public class CommonUtil {
+
+    private static final Logger logger = LoggerFactory.getLogger(CommonUtil.class);
+
+    private Map itemStatusMap;
+    private Map collectionGroupMap;
+    private Map institutionEntityMap;
+
+    @Autowired
+    private ItemStatusDetailsRepository itemStatusDetailsRepository;
+
+    @Autowired
+    private CollectionGroupDetailsRepository collectionGroupDetailsRepository;
+
+    @Autowired
+    private InstitutionDetailsRepository institutionDetailsRepository;
+
+    @Autowired
+    private ItemDetailsRepository itemDetailsRepository;
+
+    @Autowired
+    ItemChangeLogDetailsRepository itemChangeLogDetailsRepository;
 
     /**
      * This method builds Holdings Entity from holdings content
@@ -50,6 +73,16 @@ public class CommonUtil {
         failureSubmitCollectionReportInfoList.add(submitCollectionReportInfo);
     }
 
+    public void buildSubmitCollectionReportInfoWhenNoGroupIdAndAddFailures(BibliographicEntity incomingBibliographicEntity, List<SubmitCollectionReportInfo> failureSubmitCollectionReportInfoList, String owningInstitution, ItemEntity incomingItemEntity) {
+        SubmitCollectionReportInfo submitCollectionReportInfo = new SubmitCollectionReportInfo();
+        submitCollectionReportInfo.setItemBarcode(incomingItemEntity.getBarcode());
+        submitCollectionReportInfo.setCustomerCode(incomingItemEntity.getCustomerCode());
+        submitCollectionReportInfo.setOwningInstitution(owningInstitution);
+        submitCollectionReportInfo.setMessage(RecapConstants.SUBMIT_COLLECTION_FAILED_RECORD+" - "+"Unable to update dummy record, CGD is unavailable in the incoming xml record - incoming owning institution bib id - "+incomingBibliographicEntity.getOwningInstitutionBibId()
+                +", incoming owning institution item id - "+incomingItemEntity.getOwningInstitutionItemId());
+        failureSubmitCollectionReportInfoList.add(submitCollectionReportInfo);
+    }
+
     private String getExistingItemEntityOwningInstItemId(BibliographicEntity fetchedBibliographicEntity,ItemEntity incomingItemEntity){
         for(ItemEntity fetchedItemEntity:fetchedBibliographicEntity.getItemEntities()){
             if(fetchedItemEntity.getOwningInstitutionItemId().equals(incomingItemEntity.getOwningInstitutionItemId())){
@@ -57,5 +90,128 @@ public class CommonUtil {
             }
         }
         return "";
+    }
+
+    public void addItemAndReportEntities(List<ItemEntity> itemEntities, List<ReportEntity> reportEntities, boolean processHoldings, HoldingsEntity holdingsEntity, Map<String, Object> itemMap) {
+        ItemEntity itemEntity = (ItemEntity) itemMap.get("itemEntity");
+        ReportEntity itemReportEntity = (ReportEntity) itemMap.get("itemReportEntity");
+        if (itemReportEntity != null) {
+            reportEntities.add(itemReportEntity);
+        } else if (processHoldings) {
+            if (holdingsEntity.getItemEntities() == null) {
+                holdingsEntity.setItemEntities(new ArrayList<>());
+            }
+            holdingsEntity.getItemEntities().add(itemEntity);
+            itemEntities.add(itemEntity);
+        }
+    }
+
+    /**
+     * Gets item status map.
+     *
+     * @return the item status map
+     */
+    public Map getItemStatusMap() {
+        if (null == itemStatusMap) {
+            itemStatusMap = new HashMap();
+            try {
+                Iterable<ItemStatusEntity> itemStatusEntities = itemStatusDetailsRepository.findAll();
+                for (Iterator iterator = itemStatusEntities.iterator(); iterator.hasNext(); ) {
+                    ItemStatusEntity itemStatusEntity = (ItemStatusEntity) iterator.next();
+                    itemStatusMap.put(itemStatusEntity.getStatusCode(), itemStatusEntity.getId());
+                }
+            } catch (Exception e) {
+                logger.error(RecapCommonConstants.LOG_ERROR,e);
+            }
+        }
+        return itemStatusMap;
+    }
+
+    /**
+     * Gets collection group map.
+     *
+     * @return the collection group map
+     */
+    public Map getCollectionGroupMap() {
+        if (null == collectionGroupMap) {
+            collectionGroupMap = new HashMap();
+            try {
+                Iterable<CollectionGroupEntity> collectionGroupEntities = collectionGroupDetailsRepository.findAll();
+                for (Iterator iterator = collectionGroupEntities.iterator(); iterator.hasNext(); ) {
+                    CollectionGroupEntity collectionGroupEntity = (CollectionGroupEntity) iterator.next();
+                    collectionGroupMap.put(collectionGroupEntity.getCollectionGroupCode(), collectionGroupEntity.getId());
+                }
+            } catch (Exception e) {
+                logger.error(RecapCommonConstants.LOG_ERROR,e);
+            }
+        }
+        return collectionGroupMap;
+    }
+
+    /**
+     * Gets institution entity map.
+     *
+     * @return the institution entity map
+     */
+    public Map getInstitutionEntityMap() {
+        if (null == institutionEntityMap) {
+            institutionEntityMap = new HashMap();
+            try {
+                Iterable<InstitutionEntity> institutionEntities = institutionDetailsRepository.findAll();
+                for (Iterator iterator = institutionEntities.iterator(); iterator.hasNext(); ) {
+                    InstitutionEntity institutionEntity = (InstitutionEntity) iterator.next();
+                    institutionEntityMap.put(institutionEntity.getInstitutionCode(), institutionEntity.getId());
+                }
+            } catch (Exception e) {
+                logger.error(RecapCommonConstants.LOG_ERROR,e);
+            }
+        }
+        return institutionEntityMap;
+    }
+
+    /**
+     * Rollback update item availabiluty status.
+     *
+     * @param itemEntity the item entity
+     * @param userName   the user name
+     */
+    public void rollbackUpdateItemAvailabilutyStatus(ItemEntity itemEntity, String userName) {
+        ItemStatusEntity itemStatusEntity = itemStatusDetailsRepository.findByStatusCode(RecapCommonConstants.AVAILABLE);
+        itemEntity.setItemAvailabilityStatusId(itemStatusEntity.getId()); // Available
+        itemEntity.setLastUpdatedBy(getUser(userName));
+        itemDetailsRepository.save(itemEntity);
+        saveItemChangeLogEntity(itemEntity.getItemId(), getUser(userName), RecapConstants.REQUEST_ITEM_AVAILABILITY_STATUS_UPDATE, RecapConstants.REQUEST_ITEM_AVAILABILITY_STATUS_DATA_ROLLBACK);
+    }
+
+    /**
+     * Save item change log entity.
+     *
+     * @param recordId      the record id
+     * @param userName      the user name
+     * @param operationType the operation type
+     * @param notes         the notes
+     */
+    public void saveItemChangeLogEntity(Integer recordId, String userName, String operationType, String notes) {
+        ItemChangeLogEntity itemChangeLogEntity = new ItemChangeLogEntity();
+        itemChangeLogEntity.setUpdatedBy(userName);
+        itemChangeLogEntity.setUpdatedDate(new Date());
+        itemChangeLogEntity.setOperationType(operationType);
+        itemChangeLogEntity.setRecordId(recordId);
+        itemChangeLogEntity.setNotes(notes);
+        itemChangeLogDetailsRepository.save(itemChangeLogEntity);
+    }
+
+    /**
+     * Gets user.
+     *
+     * @param userId the user id
+     * @return the user
+     */
+    public String getUser(String userId) {
+        if (StringUtils.isBlank(userId)) {
+            return "Discovery";
+        } else {
+            return userId;
+        }
     }
 }
